refactor package name in module model
